GS_sample/GS_requirements.txt
numpy == 1.26.4
faiss-cpu == 1.8.0 
torch == 2.4.0
nltk == 3.8.1
tqdm == 4.66.4 
rank-bm25 == 0.2.2
transformers == 4.43.3
openai == 1.37.1
pandas == 2.2.2

